test(chat): split long-session panel coverage
- keep each ChatPanel test module below the repository size target
- isolate command routing, render-window, scroll, and scheduler coverage
- clean up mounted wrappers and stub command inventories for hermetic runs
